Move PacMan.java
